+++ src/Resources/app/administration/src/app/index.js
@@ -1,0 +1,2 @@
+import './components/heptacom-admin-open-auth-scope-field';
+import './components/heptacom-admin-open-auth-user-confirm-login';