+++ src/Resources/app/administration/src/provider/jira/components/provider-settings/index.js
@@ -1,0 +1,13 @@
+import template from './provider-settings.html.twig';
+
+const { Component } = Shopware;
+
+Component.register('heptacom-admin-open-auth-provider-jira-settings', {
+    template,
+
+    props: {
+        item: {
+            required: true,
+        },
+    },
+});