+++ src/Resources/app/administration/src/provider/open_id_connect/components/provider-settings/index.js
@@ -1,0 +1,21 @@
+import template from './provider-settings.html.twig';
+
+const { Component } = Shopware;
+
+Component.register('heptacom-admin-open-auth-provider-open-id-connect-settings', {
+    template,
+
+    props: {
+        item: {
+            required: true,
+        },
+    },
+
+    watch: {
+        item(newValue) {
+            if (!newValue.config.scopes) {
+                newValue.config.scopes = [];
+            }
+        }
+    }
+});